haas_websocket/cloud_functions/haas_monitoring/requirements.txt
requests==2.28.1
google-cloud-storage==2.4.0
google-cloud-secret-manager==2.12.0
google-cloud-bigquery==3.3.6
orjson==3.7.11